Control UI view access based on a single function-specific scope
While the API does provide admin-override routes for certain functions (such as record creation), it just overcomplicates UI code to cater for both a function-specific scope and an admin scope. It is reasonable (and safe, since the backend has the final say anyway) to assume that if a role grants, say, the 'odp.record:admin' scope, it would also grant all the 'odp.record:*' function-specific scopes too. Therefore, we can enable/disable UI capabilities based on the function-specific scopes alone.