Use the URL hash for tab activation
Previously we were activating tabs via a query param, which got messy because you can't update the query on the client side, and a page reload could then land the user on an unexpected tab.

Now we update the URL hash on tab change in the client, and activate the applicable tab from the hash (which can be set on the server via _anchor) on page load.