Add package submit and cancel routes
As part of this change, we add metadata fields to package. Package metadata is generated by translation of package tags (added via UI) in the submit route. But these fields can additionally support the use case of an automated data provider supplying a complete metadata object as part of package creation, instead of having to go via the record API. The package API thus becomes a universal entry point for data submissions, while the record API is concerned solely with data curation & publication.

The thinking originally was that package tags would be translated directly into a metadata record, but retaining generated package metadata provides a clearer separation of submission and publication workflows.

Note that the jschon-translation library needs additional work to support what we are doing here.